fix: middeware
At first the middleware does its job in checking if there is a token and redirecting the user to the login page.
After the login it redirects the user directly to the index. What I did was I added a few lines of code in the case
that the user clicks an specific product and they are not logged in. After logging in, they should ideally be redirected
product page they originally clicked on.